refactor: moved logic out of App.kt
