+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/App.kt
@@ -1,0 +1,125 @@
+package no.nav.emottak.ebms
+
+import dev.reformator.stacktracedecoroutinator.runtime.DecoroutinatorRuntime
+import io.ktor.http.Headers
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.call
+import io.ktor.server.application.createRouteScopedPlugin
+import io.ktor.server.application.install
+import io.ktor.server.engine.embeddedServer
+import io.ktor.server.metrics.micrometer.MicrometerMetrics
+import io.ktor.server.netty.Netty
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.server.request.uri
+import io.ktor.server.response.respondText
+import io.ktor.server.routing.get
+import io.ktor.server.routing.routing
+import io.ktor.util.logging.KtorSimpleLogger
+import io.micrometer.prometheus.PrometheusConfig
+import io.micrometer.prometheus.PrometheusMeterRegistry
+import net.logstash.logback.marker.Markers
+import no.nav.emottak.constants.SMTPHeaders
+import no.nav.emottak.ebms.processing.ProcessingService
+import no.nav.emottak.ebms.sendin.SendInService
+import no.nav.emottak.ebms.validation.DokumentValidator
+import no.nav.emottak.util.getEnvVar
+import org.slf4j.LoggerFactory
+import java.time.Duration
+import java.time.Instant
+import kotlin.time.toKotlinDuration
+
+val log = LoggerFactory.getLogger("no.nav.emottak.ebms.App")
+
+fun main() {
+    // val database = Database(mapHikariConfig(DatabaseConfig()))
+    // database.migrate()
+    System.setProperty("io.ktor.http.content.multipart.skipTempFile", "true")
+    if (getEnvVar("NAIS_CLUSTER_NAME", "local") != "prod-fss") {
+        DecoroutinatorRuntime.load()
+    }
+    embeddedServer(Netty, port = 8080, module = Application::ebmsProviderModule, configure = {
+        this.maxChunkSize = 100000
+    }).start(wait = true)
+}
+
+fun Application.ebmsProviderModule() {
+    val cpaClient = CpaRepoClient(defaultHttpClient())
+    val validator = DokumentValidator(cpaClient)
+
+    val processingClient = PayloadProcessingClient(scopedAuthHttpClient(EBMS_PAYLOAD_SCOPE))
+    val processing = ProcessingService(processingClient)
+
+    val sendInClient = SendInClient(scopedAuthHttpClient(EBMS_SEND_IN_SCOPE))
+    val sendInService = SendInService(sendInClient)
+
+    val appMicrometerRegistry = PrometheusMeterRegistry(PrometheusConfig.DEFAULT)
+    installMicrometerRegistry(appMicrometerRegistry)
+    installRequestTimerPlugin()
+
+    install(ContentNegotiation) {
+        json()
+    }
+
+    routing {
+        get("/") {
+            call.respondText("Hello, world!")
+        }
+
+        registerHealthEndpoints(appMicrometerRegistry)
+        navCheckStatus()
+        postEbmsAsync(validator, processing)
+        postEbmsSync(validator, processing, sendInService)
+    }
+}
+
+private fun Application.installMicrometerRegistry(appMicrometerRegistry: PrometheusMeterRegistry) {
+    install(MicrometerMetrics) {
+        registry = appMicrometerRegistry
+    }
+}
+
+private fun Application.installRequestTimerPlugin() {
+    install(
+        createRouteScopedPlugin("RequestTimer") {
+            val simpleLogger = KtorSimpleLogger("RequestTimerLogger")
+            val timeableURIs = listOf("/ebms/sync")
+            var startTime = Instant.now()
+            onCall { call ->
+                if (call.request.uri in timeableURIs) {
+                    startTime = Instant.now()
+                    simpleLogger.info("Received " + call.request.uri)
+                }
+            }
+            onCallRespond { call ->
+                if (call.request.uri in timeableURIs) {
+                    val endTime = Duration.between(
+                        startTime,
+                        Instant.now()
+                    )
+                    simpleLogger.info(
+                        Markers.appendEntries(
+                            mapOf(
+                                Pair("Headers", call.request.headers.actuallyUsefulToString()),
+                                Pair("smtpMessageId", call.request.headers[SMTPHeaders.MESSAGE_ID] ?: "-"),
+                                Pair("Endpoint", call.request.uri),
+                                Pair("request_time", endTime.toMillis()),
+                                Pair("httpStatus", call.response.status()?.value ?: 0)
+                            )
+                        ),
+                        "Finished " + call.request.uri + " request. Processing time: " + endTime.toKotlinDuration()
+                    )
+                }
+            }
+        }
+    )
+}
+
+private fun Headers.actuallyUsefulToString(): String {
+    val sb = StringBuilder()
+    entries().forEach {
+        sb.append(it.key).append(":")
+            .append("[${it.value.joinToString()}]\n")
+    }
+    return sb.toString()
+}
