+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/Routes.kt
@@ -13,8 +13,6 @@ import io.ktor.server.routing.post
 import io.micrometer.prometheus.PrometheusMeterRegistry
 import kotlinx.serialization.Serializable
 import no.nav.emottak.constants.SMTPHeaders
-import no.nav.emottak.ebms.ebxml.log
-import no.nav.emottak.ebms.ebxml.logger
 import no.nav.emottak.ebms.ebxml.receiveEbmsDokument
 import no.nav.emottak.ebms.ebxml.respondEbmsDokument
 import no.nav.emottak.ebms.model.EbMSDocument
@@ -34,11 +32,7 @@ import no.nav.emottak.melding.model.PayloadProcessing
 import no.nav.emottak.melding.model.SignatureDetails
 import no.nav.emottak.util.marker
 import no.nav.emottak.util.retrieveLoggableHeaderPairs
-import java.util.*
-import no.nav.emottak.ebms.ebxml.log
-import no.nav.emottak.ebms.ebxml.logger
-import no.nav.emottak.ebms.ebxml.receiveEbmsDokument
-import no.nav.emottak.ebms.ebxml.respondEbmsDokument
+import java.util.UUID
 
 fun Route.postEbmsSync(
     validator: DokumentValidator,
@@ -55,7 +49,7 @@ fun Route.postEbmsSync(
         ebMSDocument = call.receiveEbmsDokument()
         log.info(ebMSDocument.messageHeader().marker(loggableHeaders), "Melding mottatt")
     } catch (ex: MimeValidationException) {
-        logger().error(
+        log.error(
             call.request.headers.marker(),
             "Mime validation has failed: ${ex.message} Message-Id ${call.request.header(SMTPHeaders.MESSAGE_ID)}",
             ex
@@ -63,7 +57,7 @@ fun Route.postEbmsSync(
         call.respond(HttpStatusCode.InternalServerError, ex.parseAsSoapFault())
         return@post
     } catch (ex: Exception) {
-        logger().error(
+        log.error(
             call.request.headers.marker(),
             "Unable to transform request into EbmsDokument: ${ex.message} " +
                 "Message-Id ${call.request.header(SMTPHeaders.MESSAGE_ID)}",
@@ -146,14 +140,14 @@ fun Route.postEbmsAsync(validator: DokumentValidator, processingService: Process
             call.request.validateMime()
             ebMSDocument = call.receiveEbmsDokument()
         } catch (ex: MimeValidationException) {
-            logger().error(
+            log.error(
                 "Mime validation has failed: ${ex.message} Message-Id ${call.request.header(SMTPHeaders.MESSAGE_ID)}",
                 ex
             )
             call.respond(HttpStatusCode.InternalServerError, ex.parseAsSoapFault())
             return@post
         } catch (ex: Exception) {
-            logger().error(
+            log.error(
                 "Unable to transform request into EbmsDokument: ${ex.message} Message-Id ${
                 call.request.header(
                     SMTPHeaders.MESSAGE_ID
