+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/ebxml/ApplicationCallExt.kt
@@ -1,59 +1,25 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
 package no.nav.emottak.ebms.ebxml
 
-import dev.reformator.stacktracedecoroutinator.runtime.DecoroutinatorRuntime
 import io.ktor.client.request.forms.MultiPartFormDataContent
 import io.ktor.http.ContentType
-import io.ktor.http.Headers
 import io.ktor.http.HeadersBuilder
 import io.ktor.http.HttpStatusCode
 import io.ktor.http.content.PartData
 import io.ktor.http.content.forEachPart
 import io.ktor.http.content.streamProvider
-import io.ktor.serialization.kotlinx.json.json
-import io.ktor.server.application.Application
 import io.ktor.server.application.ApplicationCall
-import io.ktor.server.application.call
-import io.ktor.server.application.createRouteScopedPlugin
-import io.ktor.server.application.install
-import io.ktor.server.engine.embeddedServer
-import io.ktor.server.metrics.micrometer.MicrometerMetrics
-import io.ktor.server.netty.Netty
-import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.server.request.contentType
 import io.ktor.server.request.header
 import io.ktor.server.request.receive
 import io.ktor.server.request.receiveMultipart
-import io.ktor.server.request.uri
 import io.ktor.server.response.respond
 import io.ktor.server.response.respondText
-import io.ktor.server.routing.get
-import io.ktor.server.routing.routing
-import io.ktor.util.logging.KtorSimpleLogger
-import io.micrometer.prometheus.PrometheusConfig
-import io.micrometer.prometheus.PrometheusMeterRegistry
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import net.logstash.logback.marker.Markers
 import no.nav.emottak.constants.SMTPHeaders
-import no.nav.emottak.ebms.CpaRepoClient
-import no.nav.emottak.ebms.EBMS_PAYLOAD_SCOPE
-import no.nav.emottak.ebms.EBMS_SEND_IN_SCOPE
-import no.nav.emottak.ebms.PayloadProcessingClient
-import no.nav.emottak.ebms.SendInClient
-import no.nav.emottak.ebms.defaultHttpClient
+import no.nav.emottak.ebms.log
 import no.nav.emottak.ebms.model.DokumentType
 import no.nav.emottak.ebms.model.EbMSDocument
-import no.nav.emottak.ebms.navCheckStatus
-import no.nav.emottak.ebms.postEbmsAsync
-import no.nav.emottak.ebms.postEbmsSync
-import no.nav.emottak.ebms.processing.ProcessingService
-import no.nav.emottak.ebms.registerHealthEndpoints
-import no.nav.emottak.ebms.scopedAuthHttpClient
-import no.nav.emottak.ebms.sendin.SendInService
-import no.nav.emottak.ebms.validation.DokumentValidator
 import no.nav.emottak.ebms.validation.MimeHeaders
 import no.nav.emottak.ebms.validation.MimeValidationException
 import no.nav.emottak.ebms.validation.validateMimeAttachment
@@ -63,146 +29,10 @@ import no.nav.emottak.ebms.xml.asString
 import no.nav.emottak.ebms.xml.getDocumentBuilder
 import no.nav.emottak.melding.model.EbmsAttachment
 import no.nav.emottak.util.createUniqueMimeMessageId
-import no.nav.emottak.util.getEnvVar
-import org.slf4j.LoggerFactory
 import java.io.ByteArrayInputStream
 import java.nio.charset.StandardCharsets
-import java.time.Duration
-import java.time.Instant
 import java.util.Base64
 import java.util.UUID
-import kotlin.time.toKotlinDuration
-import no.nav.emottak.ebms.CpaRepoClient
-import no.nav.emottak.ebms.EBMS_PAYLOAD_SCOPE
-import no.nav.emottak.ebms.EBMS_SEND_IN_SCOPE
-import no.nav.emottak.ebms.PayloadProcessingClient
-import no.nav.emottak.ebms.SendInClient
-import no.nav.emottak.ebms.defaultHttpClient
-import no.nav.emottak.ebms.navCheckStatus
-import no.nav.emottak.ebms.postEbmsAsync
-import no.nav.emottak.ebms.postEbmsSync
-import no.nav.emottak.ebms.registerHealthEndpoints
-import no.nav.emottak.ebms.scopedAuthHttpClient
-
-val log = LoggerFactory.getLogger("no.nav.emottak.ebms.App")
-
-fun logger() = log
-fun main() {
-    // val database = Database(mapHikariConfig(DatabaseConfig()))
-    // database.migrate()
-    System.setProperty("io.ktor.http.content.multipart.skipTempFile", "true")
-    if (getEnvVar("NAIS_CLUSTER_NAME", "local") != "prod-fss") {
-        DecoroutinatorRuntime.load()
-    }
-    embeddedServer(Netty, port = 8080, module = Application::ebmsProviderModule, configure = {
-        this.maxChunkSize = 100000
-    }).start(wait = true)
-}
-
-fun PartData.payload(clearText: Boolean = false): ByteArray {
-    return when (this) {
-        is PartData.FormItem -> if (clearText) {
-            return this.value.toByteArray()
-        } else {
-            try {
-                Base64.getMimeDecoder().decode(this.value.trim())
-            } catch (e: IllegalArgumentException) {
-                log.warn("First characters in failing string: <${this.value.substring(0,50)}>", e)
-                log.warn("Last characters in failing string: <${this.value.takeLast(50)}>", e)
-                throw e
-            }
-        }
-
-        is PartData.FileItem -> {
-            val stream = this.streamProvider.invoke()
-            val bytes = stream.readAllBytes()
-            stream.close()
-            if (clearText) return bytes else Base64.getMimeDecoder().decode(bytes)
-        }
-        else -> byteArrayOf()
-    }
-}
-
-fun Application.ebmsProviderModule() {
-    val cpaClient = CpaRepoClient(defaultHttpClient())
-    val validator = DokumentValidator(cpaClient)
-
-    val processingClient = PayloadProcessingClient(scopedAuthHttpClient(EBMS_PAYLOAD_SCOPE))
-    val processing = ProcessingService(processingClient)
-
-    val sendInClient = SendInClient(scopedAuthHttpClient(EBMS_SEND_IN_SCOPE))
-    val sendInService = SendInService(sendInClient)
-
-    val appMicrometerRegistry = PrometheusMeterRegistry(PrometheusConfig.DEFAULT)
-    installMicrometerRegistry(appMicrometerRegistry)
-    installRequestTimerPlugin()
-
-    install(ContentNegotiation) {
-        json()
-    }
-
-    routing {
-        get("/") {
-            call.respondText("Hello, world!")
-        }
-
-        registerHealthEndpoints(appMicrometerRegistry)
-        navCheckStatus()
-        postEbmsAsync(validator, processing)
-        postEbmsSync(validator, processing, sendInService)
-    }
-}
-
-private fun Application.installMicrometerRegistry(appMicrometerRegistry: PrometheusMeterRegistry) {
-    install(MicrometerMetrics) {
-        registry = appMicrometerRegistry
-    }
-}
-
-private fun Application.installRequestTimerPlugin() {
-    install(
-        createRouteScopedPlugin("RequestTimer") {
-            val simpleLogger = KtorSimpleLogger("RequestTimerLogger")
-            val timeableURIs = listOf("/ebms/sync")
-            var startTime = Instant.now()
-            onCall { call ->
-                if (call.request.uri in timeableURIs) {
-                    startTime = Instant.now()
-                    simpleLogger.info("Received " + call.request.uri)
-                }
-            }
-            onCallRespond { call ->
-                if (call.request.uri in timeableURIs) {
-                    val endTime = Duration.between(
-                        startTime,
-                        Instant.now()
-                    )
-                    simpleLogger.info(
-                        Markers.appendEntries(
-                            mapOf(
-                                Pair("Headers", call.request.headers.actuallyUsefulToString()),
-                                Pair("smtpMessageId", call.request.headers[SMTPHeaders.MESSAGE_ID] ?: "-"),
-                                Pair("Endpoint", call.request.uri),
-                                Pair("request_time", endTime.toMillis()),
-                                Pair("httpStatus", call.response.status()?.value ?: 0)
-                            )
-                        ),
-                        "Finished " + call.request.uri + " request. Processing time: " + endTime.toKotlinDuration()
-                    )
-                }
-            }
-        }
-    )
-}
-
-fun Headers.actuallyUsefulToString(): String {
-    val sb = StringBuilder()
-    entries().forEach {
-        sb.append(it.key).append(":")
-            .append("[${it.value.joinToString()}]\n")
-    }
-    return sb.toString()
-}
 
 @Throws(MimeValidationException::class)
 suspend fun ApplicationCall.receiveEbmsDokument(): EbMSDocument {
@@ -217,8 +47,9 @@ suspend fun ApplicationCall.receiveEbmsDokument(): EbMSDocument {
                     if (it is PartData.FormItem) {
                         val boundary = this@receiveEbmsDokument.request.contentType().parameter("boundary")
                         if (it.value.contains("--$boundary--")) {
-                            logger().warn("Encountered KTOR bug, trimming boundary")
-                            partDataToAdd = PartData.FormItem(it.value.substringBefore("--$boundary--").trim(), {}, it.headers)
+                            log.warn("Encountered KTOR bug, trimming boundary")
+                            partDataToAdd =
+                                PartData.FormItem(it.value.substringBefore("--$boundary--").trim(), {}, it.headers)
                         }
                     }
                     this.add(partDataToAdd)
@@ -227,20 +58,24 @@ suspend fun ApplicationCall.receiveEbmsDokument(): EbMSDocument {
                 }
             }
             val start = contentType.parameter("start") ?: allParts.first().headers[MimeHeaders.CONTENT_ID]
+
             val dokument = allParts.find {
                 it.headers[MimeHeaders.CONTENT_ID] == start
             }!!.also {
                 it.validateMimeSoapEnvelope()
             }.let {
-                val contentID = it.headers[MimeHeaders.CONTENT_ID]?.convertToValidatedContentID() ?: "GENERERT-${UUID.randomUUID()}"
+                val contentID =
+                    it.headers[MimeHeaders.CONTENT_ID]?.convertToValidatedContentID() ?: "GENERERT-${UUID.randomUUID()}"
                 val isBase64 = "base64".equals(it.headers[MimeHeaders.CONTENT_TRANSFER_ENCODING], true)
                 Pair(contentID, it.payload(debugClearText || !isBase64))
             }
+
             val attachments =
                 allParts.filter { it.headers[MimeHeaders.CONTENT_ID] != start }
             attachments.forEach {
                 it.validateMimeAttachment()
             }
+
             EbMSDocument(
                 dokument.first,
                 getDocumentBuilder().parse(ByteArrayInputStream(dokument.second)),
@@ -332,5 +167,30 @@ suspend fun ApplicationCall.respondEbmsDokument(ebmsDokument: EbMSDocument) {
         this.response.headers.append(MimeHeaders.CONTENT_TYPE, ContentType.Text.Xml.toString())
         this.response.headers.append(MimeHeaders.CONTENT_TRANSFER_ENCODING, "8bit")
         this.respondText(status = HttpStatusCode.OK, text = ebmsDokument.dokument.asString())
+    }
+}
+
+private fun PartData.payload(clearText: Boolean = false): ByteArray {
+    return when (this) {
+        is PartData.FormItem -> if (clearText) {
+            return this.value.toByteArray()
+        } else {
+            try {
+                Base64.getMimeDecoder().decode(this.value.trim())
+            } catch (e: IllegalArgumentException) {
+                log.warn("First characters in failing string: <${this.value.substring(0, 50)}>", e)
+                log.warn("Last characters in failing string: <${this.value.takeLast(50)}>", e)
+                throw e
+            }
+        }
+
+        is PartData.FileItem -> {
+            val stream = this.streamProvider.invoke()
+            val bytes = stream.readAllBytes()
+            stream.close()
+            if (clearText) return bytes else Base64.getMimeDecoder().decode(bytes)
+        }
+
+        else -> byteArrayOf()
     }
 }
