+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/processing/ProcessingService.kt
@@ -6,7 +6,7 @@ import io.ktor.http.HttpStatusCode
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import no.nav.emottak.ebms.PayloadProcessingClient
-import no.nav.emottak.ebms.ebxml.logger
+import no.nav.emottak.ebms.log
 import no.nav.emottak.ebms.model.Acknowledgment
 import no.nav.emottak.ebms.model.EbmsFail
 import no.nav.emottak.ebms.model.EbmsMessage
@@ -43,7 +43,7 @@ class ProcessingService(private val httpClient: PayloadProcessingClient) {
                 direction
             )
         } catch (clientRequestException: ClientRequestException) {
-            logger().error(
+            log.error(
                 payloadMessage.marker(),
                 "Processing failed: ${clientRequestException.message}",
                 clientRequestException
